+++ arviz/data/inference_data.py
@@ -9,6 +9,7 @@ from copy import copy as ccopy
 from copy import deepcopy
 from datetime import datetime
 from html import escape
+import re
 from typing import (
     TYPE_CHECKING,
     Any,
@@ -314,7 +315,7 @@ class InferenceData(Mapping[str, xr.Dataset]):
         return InferenceData.InferenceDataItemsView(self)
 
     @staticmethod
-    def from_netcdf(filename: str) -> "InferenceData":
+    def from_netcdf(filename, group_kwargs=None, regex=False) -> "InferenceData":
         """Initialize object from a netcdf file.
 
         Expects that the file will have groups, each of which can be loaded by xarray.
@@ -326,23 +327,41 @@ class InferenceData(Mapping[str, xr.Dataset]):
         ----------
         filename : str
             location of netcdf file
+        group_kwargs : dict of {str: dict}, optional
+            Keyword arguments to be passed into each call of :func:`xarray.open_dataset`.
+            The keys of the higher level should be group names or regex matching group
+            names, the inner dicts re passed to ``open_dataset``
+            This feature is currently experimental.
+        regex : bool, default False
+            Specifies where regex search should be used to extend the keyword arguments.
+            This feature is currently experimental.
 
         Returns
         -------
         InferenceData object
         """
         groups = {}
+
         try:
             with nc.Dataset(filename, mode="r") as data:
                 data_groups = list(data.groups)
 
             for group in data_groups:
-                with xr.open_dataset(filename, group=group) as data:
+
+                group_kws = {}
+                if group_kwargs is not None and regex is False:
+                    group_kws = group_kwargs.get(group, {})
+                if group_kwargs is not None and regex is True:
+                    for key, kws in group_kwargs.items():
+                        if re.search(key, group):
+                            group_kws = kws
+                with xr.open_dataset(filename, group=group, **group_kws) as data:
                     if rcParams["data.load"] == "eager":
                         groups[group] = data.load()
                     else:
                         groups[group] = data
-            return InferenceData(**groups)
+            res = InferenceData(**groups)
+            return res
         except OSError as e:  # pylint: disable=invalid-name
             if e.errno == -101:
                 raise type(e)(